Resource selection pane
Added skeleton for resource selection pane. Player can now press F to open service menu, select the service category, select the service, select the resource, and see details about the resource type.

Also merged the Simulation menu and level scenes, because the UI had to be moved to camera space for example resource models to work properly in the UI.